move UnauthorizedException to internal.exception package
